Сделал интерфейс для создания группы и курса
